30/1/2026 5:30

+++ database/migrations/v4_3_1/master/2025_12_30_100654_add_new_col_company_details_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('company_details', function (Blueprint $table) {
-            $table->bigInteger('alternative_number')->nullable();
+            $table->integer('alternative_number')->nullable();
         });
     }
